c++: Default Arguments in Functions
Parameters without default arguments come first. This will work for add_nums(3); because the computer knows 3 corresponds to num1:

int add_nums(int num1, int num2 = 0);

But the following does NOT work for add_num(3); the computer assumes that 3 still corresponds to num1:

int add_nums(int num1 = 0, int num2);

Similarly, when a function has two default arguments, you still need to call with both arguments — if BOTH of the following are true:

The first argument IS the default value.
The second argument is NOT the default value.

You’re automating a coffee delivery system where people can order:

Coffee black (with no milk or sugar)
Coffee with milk
Coffee with sugar
Coffee with milk and sugar
It’s a bunch of options, but most of the time, people just want plain black coffee. You have a function make_coffee() built out, but it’s not efficient for people to always enter that they are NOT having milk or sugar… So what are you to do?

So, in the program, default arguments were false. 
That's why in first output we passed nothing. Then in second output the first parameter was meant to be true so we defined only one argument as the compiler passes the value to the first parameter by default. Then in third, both were meant to be true. Concentrate on Last output. We didn't used one argument as the compiler don't know that it's for which parameter and it will send that to the parameter written first. That's why we used false and true combo.