c++: Function Templates
Templates let us choose the type implementation right when you call the function. The type we choose may apply to the return type, a parameter type, or both.

Note: Using templates will slow down the program’s compile time, but speed up the execution time.

When the compiler encounters this call to a template function, it uses the template to automatically generate a function replacing each appearance of T by the type passed as the actual template parameter (int or double or etc.) and then calls it. This process is automatically performed by the compiler and is invisible to the programmer.