c++: Access Control: Public and Private [Classes]
By default, everything in a class is private, meaning class members are limited to the scope of the class. This makes it easier to keep data from mistakenly being altered, and abstracts away all the nitty gritty details.
If we will try to access the variable directly, it will throw the error.
For Ex: electric_relaxation.artist = "A Tribe Called Quest";
We added this line in the program. It will throw the error. (error: 'std::__cxx11::string Song::artist' is private within this context
   electric_relaxation.artist = "A Tribe Called Quest";)

To Solve this Problem and use the private variable we did the following:

   A better solution that allows you to interact with artist indirectly:

   * Create a public .add_artist() for setting the artist.
   * Create a public .get_artist() for retrieving the artist.

   In main(), set electric_relaxation‘s artist to "A Tribe Called Quest" using .add_artist().

  And it the end retrieve and print out the electric_relaxation‘s artist using .get_artist().