C++: Destructor
A destructor is a special method that handles object destruction. Like a constructor, it has the same name as the class and no return type, but is preceded by a ~ operator and takes no parameters.
Destructor is basically a member function which destructs or deletes an object.
You generally won’t need to call a destructor; the destructor will be called automatically in any of the following scenarios:

(1) the object goes out of scope
(2) the function ends
(3) the program ends
(4) a block containing local variables ends
(5) a delete operator is called

The Syntax to define a Destructor in Class is:
~Class_name();

& The Syntax of a Destructor is:
Class_name::~Class_name(){
  //Any final cleanup
}