c++: Range based for loop [vectors]
Syntax of Range Based for loop in case of vectors:
for (data_type variable : vector)

Here the data type is the type of data that is stored in the vector. The variable name can be anything and there is no need to define that variable. The vector passes the value of each element one by one in that variable till the end. The loop will automatically terminate when the vector will be empty.